Refactoring complete

+++ CPT Project/Assets/Scripts/Corner.cs
@@ -10,7 +10,7 @@ public class Corner : MonoBehaviour {
     #endregion
     #region Final Corner Data
     private int index;                  //corders order
-    private double radius;              //radius of corner in degrees
+    private float radius;              //radius of corner in degrees
     private bool direction;      //direction of corner (clockwise/counter clockwise)
     private Vector2 finalCoordinates;    // final coordinates of corners center     
     #endregion
@@ -18,7 +18,7 @@ public class Corner : MonoBehaviour {
     #region Getters
     public Vector2 getPlaceholderCoordinates() { return placeholderCoordinates; }
     public int getIndex() { return index; }
-    public double getRadius() { return radius; }
+    public float getRadius() { return radius; }
     public bool getDiretcion() { return direction; }
     public Vector2 getFinalCoordinates() { return finalCoordinates; }
     #endregion
@@ -26,7 +26,7 @@ public class Corner : MonoBehaviour {
     #region Setters
     public void setPlaceholderCoordinates(Vector2 a) { placeholderCoordinates = a; }
     public void setIndex(int b) { index = b; }
-    public void setRadius(double c) { radius = c; }
+    public void setRadius(float c) { radius = c; }
     public void setDirection(bool d) { direction = d; }
     public void setFinalCoordinates(Vector2 e) { finalCoordinates = e; } 
     #endregion
